@@ -24,59 +24,57 @@
         "files": [ "vendor/symfony/symfony/src/Symfony/Component/VarDumper/Resources/functions/dump.php" ]
     },
     "require": {
-        "php": "~5.6|~7.0",
-        "ext-imagick": "*",
-        "symfony/symfony": "~2.8",
-        "twig/extensions": "~1.4",
+        "php": "^7.1",
+        "symfony/symfony": "^3.4.2",
+        "twig/twig": "^1.0||^2.0",
+        "twig/extensions": "^1.5.0",
         "symfony/assetic-bundle": "~2.8",
         "symfony/swiftmailer-bundle": "~2.5",
-        "symfony/monolog-bundle": "~2.12|~3.0",
-        "sensio/distribution-bundle": "^3.0.36|^4.0.6|^5.0",
-        "sensio/generator-bundle": "^2.3|^3.0",
+        "symfony/monolog-bundle": "^3.1.0",
+        "doctrine/doctrine-bundle": "^1.6",
+        "doctrine/orm": "^2.5",
+        "sensio/distribution-bundle": "^5.0.19",
         "incenteev/composer-parameter-handler": "~2.1",
         "tedivm/stash-bundle": "~0.6",
-        "ezsystems/ezpublish-kernel": "^6.12",
-        "ezsystems/repository-forms": "^1.10",
-        "ezsystems/ezplatform-solr-search-engine": "^1.5",
-        "ezsystems/platform-ui-bundle": "^1.12",
-        "ezsystems/ez-support-tools": "~0.1.3",
-        "ezsystems/ezplatform-http-cache": "^0.2",
-        "ezplatform-i18n/ezplatform-i18n-ach_ug": "^1.4",
-        "ezsystems/ezplatform-multi-file-upload": "^0.1",
-        "ezsystems/ezplatform-design-engine": "^1.1",
-        "egulias/listeners-debug-command-bundle": "~1.9",
-        "white-october/pagerfanta-bundle": "1.0.*",
-        "hautelook/templated-uri-bundle": "~1.0 | ~2.0",
-        "doctrine/doctrine-bundle": "~1.5",
-        "sensio/framework-extra-bundle": "~3.0",
-        "ezsystems/content-on-the-fly-prototype": "~0.1.11",
+        "ezsystems/ezpublish-kernel": "~7.0.0",
+        "ezsystems/repository-forms": "~2.0.0",
+        "ezsystems/ezplatform-solr-search-engine": "~1.5.1",
+        "ezsystems/ez-support-tools": "~0.2.1",
+        "ezsystems/ezplatform-http-cache": "~0.4.0",
+        "ezsystems/ezplatform-admin-ui": "~1.0.0",
+        "ezsystems/ezplatform-admin-ui-modules": "~1.0.0",
+        "ezsystems/ezplatform-admin-ui-assets": "~1.0.0",
+        "ezsystems/ezplatform-design-engine": "~1.1.0",
+        "knplabs/knp-menu-bundle": "^2.1",
+        "sensio/framework-extra-bundle": "^3.0.2",
         "willdurand/js-translation-bundle": "^2.6.4",
+        "oneup/flysystem-bundle": "^2.0",
+        "friendsofsymfony/jsrouting-bundle": "^1.4",
+        "white-october/pagerfanta-bundle": "^1.1",
         "roave/security-advisories": "dev-master",
+        "leafo/scssphp": "^0.7",
         "ezsystems/comments-bundle": "~6.0",
         "knplabs/packagist-api": "dev-master",
         "netgen/tagsbundle": "^3.0"
     },
     "require-dev": {
-        "ezsystems/ezplatform-i18n": "^1.4",
         "behat/behat": "~3.2",
         "behat/symfony2-extension": "~2.0",
         "behat/mink-extension": "~2.0",
         "behat/mink-goutte-driver": "~1.0",
         "behat/mink-selenium2-driver": "~1.0",
-        "jarnaiz/behat-junit-formatter": "^1.3",
-        "ezsystems/behatbundle": "^6.3"
-    },
-    "suggest": {
-        "ezsystems/legacy-bridge": "Provides the full legacy backoffice and legacy features"
+        "ezsystems/behatbundle": "^6.5",
+        "phpunit/phpunit": "^6.4",
+        "sensio/generator-bundle": "~3.1",
+        "symfony/phpunit-bridge": "~3.2"
     },
     "scripts": {
         "symfony-scripts": [
             "Incenteev\\ParameterHandler\\ScriptHandler::buildParameters",
-            "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::buildBootstrap",
             "eZ\\Bundle\\EzPublishCoreBundle\\Composer\\ScriptHandler::clearCache",
             "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::installAssets",
-            "eZ\\Bundle\\EzPublishCoreBundle\\Composer\\ScriptHandler::dumpAssets",
-            "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::installRequirementsFile"
+            "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::installRequirementsFile",
+            "@php bin/console assetic:dump"
         ],
         "post-install-cmd": [
             "@symfony-scripts"
