@@ -5,6 +5,7 @@ use Symfony\Component\Debug\Debug;
 
 // Ensure UTF-8 is used in string operations
 setlocale(LC_CTYPE, 'C.UTF-8');
+require __DIR__ . '/../vendor/autoload.php';
 
 // Environment is taken from "SYMFONY_ENV" variable, if not set, defaults to "prod"
 $environment = getenv('SYMFONY_ENV');
@@ -18,26 +19,11 @@ if (($useDebugging = getenv('SYMFONY_DEBUG')) === false || $useDebugging === '')
     $useDebugging = $environment === 'dev';
 }
 
-// Depending on SYMFONY_CLASSLOADER_FILE use custom class loader, otherwise use bootstrap cache, or autoload in debug
-if ($loaderFile = getenv('SYMFONY_CLASSLOADER_FILE')) {
-    require_once $loaderFile;
-} else {
-    require_once __DIR__ . '/../app/autoload.php';
-    if (!$useDebugging) {
-        require_once __DIR__ . '/../app/bootstrap.php.cache';
-    }
-}
-
 if ($useDebugging) {
     Debug::enable();
 }
 
 $kernel = new AppKernel($environment, $useDebugging);
-
-// we don't want to use the classes cache if we are in a debug session
-if (!$useDebugging) {
-    $kernel->loadClassCache();
-}
 
 // Depending on the SYMFONY_HTTP_CACHE environment variable, tells whether the internal HTTP Cache mechanism is to be used.
 // If not set, or "", it is auto activated if _not_ in "dev" environment.
@@ -47,23 +33,19 @@ if (($useHttpCache = getenv('SYMFONY_HTTP_CACHE')) === false || $useHttpCache ==
 
 // Load HTTP Cache ...
 if ($useHttpCache) {
-    // The standard HttpCache implementation can be overridden by setting the SYMFONY_HTTP_CACHE_CLASS environment variable.
-    // NOTE: Make sure to setup composer config so it is *autoloadable*, or use "SYMFONY_CLASSLOADER_FILE" for this.
-    if ($httpCacheClass = getenv('SYMFONY_HTTP_CACHE_CLASS')) {
-        $kernel = new $httpCacheClass($kernel);
-    } else {
-        $kernel = new AppCache($kernel);
+    $kernel = new AppCache($kernel);
+
+    // When using the HttpCache, you need to call the method in your front controller instead of relying on the configuration parameter
+    Request::enableHttpMethodParameterOverride();
+
+    // If you are behind one or more trusted reverse proxies, you might want to set them in SYMFONY_TRUSTED_PROXIES environment
+    // variable in order to get correct client IP. NOTE: As per Symfony doc you will need to customize these lines for your proxy!
+    if ($trustedProxies = getenv('SYMFONY_TRUSTED_PROXIES')) {
+        Request::setTrustedProxies(explode(',', $trustedProxies), Request::HEADER_X_FORWARDED_ALL);
     }
 }
 
 $request = Request::createFromGlobals();
-
-// If you are behind one or more trusted reverse proxies, you might want to set them in SYMFONY_TRUSTED_PROXIES environment
-// variable in order to get correct client IP
-if ($trustedProxies = getenv('SYMFONY_TRUSTED_PROXIES')) {
-    Request::setTrustedProxies(explode(',', $trustedProxies));
-}
-
 $response = $kernel->handle($request);
 $response->send();
 $kernel->terminate($request, $response);
